% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_results.R
\name{beta_dist}
\alias{beta_dist}
\title{Plot the posterior distribution of the coefficients}
\usage{
beta_dist(result, title = "")
}
\arguments{
\item{result}{resulting object from mmvbvs function}

\item{title}{A string object for the title of the resulting plot}
}
\value{
ggplot object
}
\description{
Plot the posterior distribution of the coefficients
}
