% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_results.R
\name{plot_beta}
\alias{plot_beta}
\title{Plot the coefficients for each variable for each iteration of MCMC}
\usage{
plot_beta(result, title = "")
}
\arguments{
\item{result}{Output object from mmvbvs function}

\item{title}{A string object for the title of the resulting plot}
}
\value{
ggplot object
}
\description{
Plot the coefficients for each variable for each iteration of MCMC
}
