% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_results.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot the the inclusion of each variable for each MCMC iteration}
\usage{
plot_gamma(result, title = "")
}
\arguments{
\item{result}{Output object from mmvbvs function}

\item{title}{A string object for the title of the resulting plot}
}
\value{
ggplot object
}
\description{
Plot the the inclusion of each variable for each MCMC iteration
}
