% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_results.R
\name{plot_sigma}
\alias{plot_sigma}
\title{Plot the posterior mean of the covariance matrix}
\usage{
plot_sigma(result, title = "")
}
\arguments{
\item{result}{resulting object from mmvbvs function}

\item{title}{A string object for the title of the resulting plot}
}
\value{
ggplot object
}
\description{
Plot the posterior mean of the covariance matrix
}
